chore(pubsub): extract pubsub interfaces to dedicated package
Ideally the interfaces could be in the domain layer (`service`) and the
implementation in a `redis` package. However, this would imply a major
refactorisation to avoid some circular dependencies.
