+++ pkg/middleware/middleware.go
@@ -9,21 +9,21 @@ import (
 
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 
-	"github.com/instill-ai/pipeline-backend/pkg/memory"
+	"github.com/instill-ai/pipeline-backend/pkg/pubsub"
 	"github.com/instill-ai/pipeline-backend/pkg/repository"
 	"github.com/instill-ai/pipeline-backend/pkg/service"
 
 	pb "github.com/instill-ai/protogen-go/pipeline/pipeline/v1beta"
 )
 
-type fn func(*runtime.ServeMux, pb.PipelinePublicServiceClient, http.ResponseWriter, *http.Request, map[string]string, memory.EventSubscriber)
+type fn func(*runtime.ServeMux, pb.PipelinePublicServiceClient, http.ResponseWriter, *http.Request, map[string]string, pubsub.EventSubscriber)
 
 // AppendCustomHeaderMiddleware appends custom headers
 func AppendCustomHeaderMiddleware(
 	mux *runtime.ServeMux,
 	client pb.PipelinePublicServiceClient,
 	next fn,
-	sub memory.EventSubscriber,
+	sub pubsub.EventSubscriber,
 ) runtime.HandlerFunc {
 	return runtime.HandlerFunc(func(w http.ResponseWriter, r *http.Request, pathParams map[string]string) {
 		next(mux, client, w, r, pathParams, sub)
